Minor refactor + comments
